moved screen module to the  vga_display driver and added documentation
