+++ constants/src/enums.rs
@@ -32,26 +32,6 @@ pub enum PacketSize {
     Default = 0x10,
 }
 
-#[repr(u8)]
-pub enum Color {
-    Black = 0,
-    Blue = 1,
-    Green = 2,
-    Cyan = 3,
-    Red = 4,
-    Magenta = 5,
-    Brown = 6,
-    LightGray = 7,
-    DarkGray = 8,
-    LightBlue = 9,
-    LightGreen = 10,
-    LightCyan = 11,
-    LightRed = 12,
-    Pink = 13,
-    Yellow = 14,
-    White = 15,
-}
-
 #[derive(Clone)]
 pub enum PageSize {
     /// 4Kib pages
