+++ kernel/src/drivers/mod.rs
@@ -1,0 +1,2 @@
+pub mod disk;
+pub mod vga_display;
